requirements.txt
flask
joblib
pandas
openai
python-dotenv
